requirements.txt
pytorch-lightning==1.8.6
albumentations==1.3.0
gdown==4.6.0
Pillow==9.4.0
timm==0.6.12
mmcv==1.7.1
streamlit==1.17.0

